Enum scoolite::command::Statement

enum Statement {
    Insert(String),
    Select,
}

Statement is the enum that contains all of the statements for scoolite. An example of a statement is insert, it does belong to the SQL specification and it is used to add a row to a table.

Variants
Insert(String)Select
Methods

impl Statement

fn from_str(input: &str) -> Result<Box<dyn Command>, Error>

Tries to parse an &str input into a Box<Command>, if it fails it returns a "not implemented error" Error.

All of the possibilities are just the variants on the enum.

fn insert(&self, input: &str, table: &mut Table) -> Result<(), Error>

Creates a new Row based of an input &str and inserts it inside of a table. This is what get's called when something like Statement::Insert("insert 1 john john@mailbox.com").execute() happens.

fn select(&self, table: &Table) -> Result<(), Error>

Prints all Rows inside of a table. This is what get's called when something like Statement::Select.execute() happens.